split components and placed in into folder
